+++ src/main/java/ru/job4j/ood/core/ocp/two/NotificationService.java
@@ -1,4 +1,4 @@
-package ru.job4j.ood.core.coreocp.two;
+package ru.job4j.ood.core.ocp.two;
 
 public class NotificationService {
