+++ src/main/java/ru/job4j/ood/core/srp/Toy.java
@@ -1,4 +1,4 @@
-package ru.job4j.ood.core.coresrp;
+package ru.job4j.ood.core.srp;
 
 public class Toy {
